+++ tests/Unit/Public/Get-ComputerName.Tests.ps1
@@ -47,27 +47,27 @@ AfterAll {
 
 Describe 'Get-ComputerName' {
     BeforeAll {
-        $mockComputerName = 'MyComputer'
+        $mockShortComputerName = 'MyComputer'
+        $mockFqdnComputerName = 'MyComputer.domain.com'
+    }
 
-        if ($IsLinux -or $IsMacOs)
-        {
-            function hostname
-            {
-            }
-
-            Mock -CommandName 'hostname' -MockWith {
-                return $mockComputerName
-            } -ModuleName 'DscResource.Common'
-        }
-        else
-        {
-            $mockComputerName = $env:COMPUTERNAME
+    Context 'When getting computer name without FQDN switch' {
+        It 'Should return the short computer name' {
+            $result = Get-ComputerName -ErrorAction Stop
+            $result | Should -BeOfType [System.String]
+            $result | Should -Not -BeNullOrEmpty
+            # Should return the same as [System.Environment]::MachineName (short name)
+            $result | Should -Be ([System.Environment]::MachineName)
         }
     }
 
-    Context 'When getting computer name' {
-        It 'Should return the correct computer name' {
-            Get-ComputerName | Should -Be $mockComputerName
+    Context 'When getting computer name with FQDN switch' {
+        It 'Should return the FQDN when DNS resolution succeeds' {
+            $result = Get-ComputerName -FullyQualifiedDomainName -ErrorAction Stop
+            $result | Should -BeOfType [System.String]
+            $result | Should -Not -BeNullOrEmpty
+            # The result should either be FQDN (if DNS works) or short name (if DNS fails)
+            # We can't predict which without mocking, so just ensure it's not empty
         }
     }
 }